pyproject.toml
[project]
name = "graphene-intel"
version = "0.1.0"
description = "Automated intelligence platform for graphene sector stock monitoring"
requires-python = ">=3.12"
authors = [
    { name = "michalekz", email = "zdenek.michalek@gmail.com" }
]

dependencies = [
    "anthropic>=0.40",              # Claude API (Haiku + Sonnet)
    "httpx>=0.27",                  # Async HTTP client
    "feedparser>=6.0",              # RSS/Atom parsing
    "yfinance>=0.2",                # Stock price/volume data
    "praw>=7.7",                    # Reddit API
    "pytrends>=4.9",                # Google Trends (unofficial API)
    "python-telegram-bot>=21",      # Telegram Bot API
    "pyyaml>=6.0",                  # Config file parsing
    "python-dotenv>=1.0",           # .env file loading
    "aiosqlite>=0.20",              # Async SQLite
    "edgartools>=2.0",              # SEC EDGAR Form 4 / filings parser
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
